# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM1 API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/client/index": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "Список клиентов",
                "operationId": "15219952ddecc6cc905fd2cdf0825145",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/": {
            "post": {
                "tags": [
                    "client"
                ],
                "summary": "Создание клиента",
                "operationId": "e89d932fa8f76f2cf83c15123e99812d",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateClientRequest"
                },
                "responses": {
                    "201": {
                        "description": "Client resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/by-email": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "Поиск по email",
                "operationId": "ab0e5cb0a96cf67d5f60ef1f61a197c0",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateClientRequest"
                },
                "responses": {
                    "200": {
                        "description": "Client resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/update/{id}": {
            "patch": {
                "tags": [
                    "client"
                ],
                "summary": "Обновление по id",
                "operationId": "00e48b6d65e258d996d2557759d5a371",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id клиента",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateClientRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/{id}": {
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "Удаление по id",
                "operationId": "b862611ecdbc4a06fbb975a2df5f0a13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id клиента",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/force-delete/{id}": {
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "Безвозвратное удаление по id",
                "operationId": "0c63ce3b24e11e3bf6dd1493777864f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id клиента",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/manager/{id}": {
            "patch": {
                "tags": [
                    "client"
                ],
                "summary": "Изменение менеджера по id клиента",
                "operationId": "f0ec38e16f6655efe64efac8ad145b29",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "description": "id клиента",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "managerId",
                        "in": "path",
                        "description": "id клиента",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "company",
                    "source",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Adam"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "t@test.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1235464333"
                    },
                    "company": {
                        "type": "string",
                        "example": "test.co"
                    },
                    "source": {
                        "type": "string",
                        "example": "call"
                    },
                    "status": {
                        "type": "string",
                        "example": "new"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "CreateClientRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "emai",
                                "phone",
                                "company",
                                "source",
                                "manager_id",
                                "status"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Adam"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "t@test.com"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "+1998573422"
                                },
                                "company": {
                                    "type": "string",
                                    "example": "test.co"
                                },
                                "source": {
                                    "description": "Из списка: advertising, call, site",
                                    "type": "string"
                                },
                                "manager_id": {
                                    "type": "integer"
                                },
                                "status": {
                                    "description": "Из списка: new, active, archived",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateClientRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Adam",
                                    "nullable": true
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "t@test.com",
                                    "nullable": true
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "+1998573422",
                                    "nullable": true
                                },
                                "company": {
                                    "type": "string",
                                    "example": "test.co",
                                    "nullable": true
                                },
                                "source": {
                                    "description": "Из списка: advertising, call, site",
                                    "type": "string",
                                    "nullable": true
                                },
                                "manager_id": {
                                    "type": "integer",
                                    "nullable": true
                                },
                                "status": {
                                    "description": "Из списка: new, active, archived",
                                    "type": "string",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}